Move download tests to source tree
