+++ src/download/bearerToken.test.ts
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: BSD-3-Clause
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-import { getBearer } from "../../src";
+import { getBearer } from "./bearerToken";
 import { expect, default as chai } from "chai";
 import chaiAsPromised from "chai-as-promised";
 import nock from "nock";
