+++ src/download/exchangeDirectoryParser.test.ts
@@ -5,7 +5,7 @@
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
-import { extractFiles } from "../../src";
+import { extractFiles } from "./exchangeDirectoryParser";
 
 import { expect, default as chai } from "chai";
 import chaiAsPromised from "chai-as-promised";
