+++ src/download/exchangeTools.test.ts
@@ -4,11 +4,12 @@
  * SPDX-License-Identifier: BSD-3-Clause
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-import { RestApi, groupByCategory } from "../../src";
 import {
+  groupByCategory,
   removeRamlLinks,
   removeVersionSpecificInformation
-} from "../../src/download/exchangeTools";
+} from "./exchangeTools";
+import { RestApi } from "./exchangeTypes";
 
 import _ from "lodash";
